Fix file explorer issue on mobile app
There seems to be a bug in the Obsidian IOS app: when you create a
recursive directory with the Obsidian file system API, the file explorer
view does not update correctly: only the outer directory gets added
in file explorer. You have to refresh the app to update file explorer.

I found a workaround by creating each directory in the path one
at a time. This forces file explorer to update correctly.